[Backport 6X] Check delay eager free in squelch functions. (#12086)
Executor try to squelch nodes in it‘s subtree after a node returning a NULL tuple.
But sometimes squelching is not safe. If the node is not safe to squelch, the
parameter ’delayEagerFree‘ will be set to true. We should check the parameter
when we squelch the node both on 'Execxxx' function and 'ExecSquelchxxxx' function.

So add code checking 'delayEagerFree' in ExecSquelchSort and
ExecSquelchFunction.

This is a backport of commit 361f211 from master.

Co-authored-by: Jinbao Chen <cjinbao@vmware.com>