pg_dumpall: Fix dump of built-in resource groups
Currently, pg_dumpall erroneously emits CREATE RESOURCE GROUP statements
for built-in resource groups: admin_group and default_group.
Consequently, this causes pg_upgrade to fail during schema restore
(since the built-in groups already exist in the target cluster).

The bug can be traced to a string comparison being done on the quoted
version of the resource group name. This commit fixes that.

Co-authored-by: Soumyadeep Chakraborty <soumyadeep2007@gmail.com>