Update pg_hba.conf remotely on segment host
gprecoverseg appends to pg_hba.conf file on every run even if the
lines are already present, creating very large pg_hba files over time.
This commit changes that, we now check if the line is present before
adding it in.

This commit adds a new sh file to run on segments to ensure
parallel execution. Each segment receives the entries to insert for the
pg_hba.conf file and parses through them to filter out duplicates and
insert them locally.

This process filters the existing duplicates from the
pg_hba.conf file as well. The filtering is done by reducing the
whitespaces to a single space. We do preserve the formatting of the
existing lines but the new ones are all single space seperated.

The semi-duplicated create_entries code in gpmovemirrors
have been moved to the common file.

Co-authored-by: Divyesh Vanjare <vanjared@vmware.com>