Support implicit CoerceViaIO casts in Subplan test expressions in ORCA
ORCA already supported CoerceViaIO casts when they're explicit in the
query. This commit expands their usage for implicit casts and adds
handling for them in subplan test exprs during DXL to PlStmt
translation.

Also refactor code in TranslateDXLSubplanTestExprToScalar() to make it
easier to extend for other cast types; and set inputcolid and opcollid
in op_expr, which were previously left unset causing an ERROR on
execution.

This fixes issue #12133.