pg_upgrade: Detect views referencing deprecated columns
This commit introduces a pg_upgrade check to look for views referencing
deprecated columns in a 5X cluster. This is similar in spirit to the
check added earlier to detect references to deprecated tables.

This is accompanied by the 5X PR: #12168, which introduces a
pg_upgrade_support UDF to detect such views.