Orca rejects partition index if root and leaf dropped column layout mismatch
ORCA for GPDB6 and older versions assumes that the root and leaf
partitions have the same underlying column structure. That assumption
can be broken when an exchange partition with or without same dropped
columns as root is inserted into the partition table. Further
complicating the matter is that ORCA always uses the root partition to
construct index metadata. This would lead to crashes when trying to
explore an Index plan.
As part of this commit, we detect a mismatch in the index and relation
metadata, and if we find a mismatch, ORCA does not consider the index.

Below are the scenarios, that illustrate the checks added for detecting
a mismatch:

Scenario 1: When querying on leaf with the following column layout
   Root (a, dropped col, b, dropped col, c)
   Leaf (a, b, c)
   Index(a, b)

Index in MD cache is as follows:

<dxl:Index ... KeyColumns="0,2" IncludedColumns="0,2,4,5,6,7,8,9,10">
   .....
</dxl:Index>

Here index 1 and 3 map to the dropped columns and 5-10 maps to system
cols in ROOT table.
When querying leaf partition, we iterate over the `IncludedColumns` from the
INDEX (0,2,4,5,6,7,8,9,10) and fetch the NonDropped column from the
relation(leaf partition-> 0,1,2,3,4,5,6,7,8,9) at the position from Index.

Previously, `PdrgpcrIndexColumns()` would fail trying to access element
for pos=10 which lies outside the colref_array bound (array size = 10 (3
user col+7 system col)).
This commit adds the check for `ulPosNonDropped >= colref_array->Size()`
to detect for a mismatch and not choose this Index.

Scenario 2: When querying on the leaf with the following column layout
   Root (a, b, c)
   Leaf (a, dropped col, b, dropped col, c)
   Index(a, b)

Index in MD cache is as follows:

<dxl:Index ... KeyColumns="0,1" IncludedColumns="0,1,2,3,4,5,6,7,8,9">
   .....
</dxl:Index>

Here index 0-2 maps to the columns(a,b,c) and 3-9 maps to system cols in
ROOT table.
When querying leaf partition, we iterate over the IncludedColumns from the
INDEX (0,1,2,3,4,5,6,7,8,9) and fetch the NonDropped column from the
relation(leaf partition-> 0,2,4,5,6,7,8,9,10,11) at the position from Index.

Previously, this would fail when `NonDroppedColAt` tries to fetch a
non-dropped column position at position 1, that doesn't exist in the map
and crashes when trying to access a NULL pointer.
This commit updates the `NonDroppedColAt()` to return gpos::ulong_max
and check for it to detect for a mismatch and not choose this Index.

Since this code is shared between IncludedColumns and KeyColumns for the
Index, added similar tests for dropped columns between index key
columns.

Note: For EXTERNAL tables, since GPDB doesn't support creating indexes
on external tables, we do not need to worry about mismatched dropped
columns for exchange partition with an external table.

Co-authored-by: Ekta Khanna <ekhanna@vmware.com>