Mirror is up only if it has received WAL
Previously, FTS used to treat a mirror as up as soon as the walsender
entered catchup state.  This patch restricts that criterion.  A mirror
is now considered up only if at least one chunk of WAL has been
received by it.

An active replication slot is obsoleted when replication lag grows
beyond max_slot_wal_keep_size limit.  This is a point of no return,
the mirror must be destroyed and recreated using pg_basebackup
(gprecoverseg -F).  The walsender process exits when this happens.
However, the walreceiver on mirror is oblivious to this and keeps on
attempting new replication connections.

Walsender enters catchup state immediately upon receiving a connection
request.  The request fails soon after, when an attempt to locate the
requested start LSN in locally available WAL fails.  An FTS probe
within this window used to erroneously treat the mirror as up and turn
syncrep on.  This is definitely not desirable.  The patch causes FTS
to treat the mirror as down even if the walsender is found in catchup
state.

Reviewed by: Ashwin Agrawal <aashwin@vmware.com>