Speedup crash_recovery_dtm test
The test used to inject fault on QE to send error respone to QD in
finish-prepared step.  The fault was injected with
gp_inject_fault_infinite interface.  It causes the fault to continue
triggerring until reset.  The test is not aware of exactly when the QD
finishes postmaster reset and start DTX recovery.  From the
coordinator logs, it was clear that the DTX recovery also hit this
fault, which was reset after a few seconds by the test. This causes
unnecessary delay.  Avoid it by disabling DTX retry logic and inject
the fault so that it is triggered at the most once time..

Reviewed by: Ashwin Agrawal, Junfeng Yang

(cherry picked from commit 20ca2441fbed6c03c45f10b80f3d3a6e831033de)