Coerce unknown-type literals to type cstring, but hide it (#12145)
Commit 951f817e9b4 "Coerce unknown-type literals to type text instead of
cstring" has an issue that it processed outfunc and infunc directly,
which made Greenplum have no chance to do an unknown-to-text-to-cstring
cast. It errors out, even panics, in some cases since the data type
don't match.

The better solution is still coercing unknown-type literals to type
cstring, but hiding it. Check out the test case to see what it fixes.