Check errno as early as possible
Previously, the result of select() system call and errno set by it was
checked after performing several function calls, including checking
for interrupts and checkForCancelFromQD.  That made it very likely for
errno to change, losing the original value that was set by the
select().

This patch fixes it so that the errno is checked immediately after the
system call.  This should address intermittent failures in CI with
error message like this:

    "ERROR","58M01","interconnect error: select: Success"

Cherry-picked from 9fd138da and follow up commits bf36fb3b and
a251127b.

Reviewed by: Paul Guo <paulguo@gmail.com>