Check lock table while cleaning resource queue lock
We introduce a validation check to PANIC when corruption of the lock
table is encountered. We also flip a log message from DEBUG1 to LOG to
make debugging easier for issues related to resource queue lock clean
up.

5X to 6X forward-port notes:
Changed validation check to check for NULL instead of INVALID_OFFSET.
This is due to changes in shared memory management from upstream
Postgres.

Postgres reference commit:
https://github.com/postgres/postgres/commit/d7112cfa884e5125a20a24590f193ceebedf0146

Co-authored-by: Yao Wang <wayao@vmware.com>
Co-authored-by: Soumyadeep Chakraborty <soumyadeep2007@gmail.com>
Co-authored-by: Ashwin Agrawal <aashwin@vmware.com>
Co-authored-by: Kate Dontsova <edontsova@pivotal.io>