Fix a bug of concurrently executing UPDATE/DELETE with VACUUM on partition ao table (#11895)
The problem is ERROR "tuple concurrently updated" occurs when running
delete qeury during ETL as below error. When we concurrently execute an
UPDATE/DELETE statement and a VACUUM statement on a partition ao table.
UPDATE/DELETE statement is likely to get an out-of-date snapshot after
the VACUUM has been processed on one of the leaf partition table.
So we fix it by acquiring locks before executing. After we get locks, we
dispatch snapshot. So this problem will not happen.

If the relation is partitioned, we should acquire corresponding locks on
each leaf partition before entering InitPlan. The snapshot will acquire
before
the InitPlan, so if we wait the locks for a while in the InitPlan. When
we
get all locks, the snapshot maybe become invalid. So we must acquire
lock
before entering InitPlan to keep the snapshot valid.
In the parse-analyze stage, we should get all locks we needed and also
need to
maintain the consistency of the locks of the parent table and the child
table.
Only one situation is special, when the command is to insert, and gdd
don't open,
the dead lock will happen.
for example:
Without locking the partition relations on QD when INSERT with Planner
the
following dead lock scenario may happen between INSERT and AppendOnly
VACUUM
drop phase on the partition table:

1. AO VACUUM drop on QD: acquired AccessExclusiveLock
2. INSERT on QE: acquired RowExclusiveLock
3. AO VACUUM drop on QE: waiting for AccessExclusiveLock
4. INSERT on QD: waiting for AccessShareLock at ExecutorEnd()

2 blocks 3, 1 blocks 4, 1 and 2 will not release their locks until 3 and
4 proceed.
Hence INSERT needs to Lock the partition tables on QD here (before 2) to
prevent
this dead lock.

This will cause a deadlock. But if gdd have already opened, gdd will
solve the
dead lock. So it is safe to not get locks for partitions.

But for the update and delete, we must acquire locks whether the gdd
opens or not.

So for the normal sql statement, we move the acquiring lock behavior in
the InitPlan
to setTargetTable. for the cached plan, we add the code acquiring locks
on the child
table.

Co-authored-by: Rui Wang wangru@vmware.com wraymo