Wrong snapshot for EOF corrupts data in AO/AOCO
In 5X a snapshot satisfies "now" was used when transaction extracted AO/AOCO
segment file eof. In other words "now" allowed us to see all committed transactions. 
When this code was ported to 6X a catalog snapshot was chosen for this purpose. 
As a result we don't see committed transactions that started after the snapshot has
been taken. As a result a transaction can change segment file eof between the
moment a snapshot was taken and an exclusive file segment lock was hold. So AO/AOCO
in 6X suffers from data corruption in concurrent DML queries.

5X and 7X (thanks to a big refactoring 5778f31124d573) do not face this problem.

This commit fixes the problem using snapshot self instead of catalog
snapshot - it allows us to see all committed data as "now" in 5X did.