Fix 6x pipeline fail (#12311)
1. skip gpload test in Red Hat for gpdb6-oracle7-test task
judge the os in gpload test Makefile, if os is red hat, than skip gpload test

2. add test case 39
case 39 is added by Zhaorui in pr Fix gpload with header and reuse_tables bug. #12166 (case number is 47 in his pr), and missed when refactor gpload test

3. delete 'ignore drop' command when compare ans and out files
drop information is ignored when comparing ans and out files in local test. While in remote test it is not ignored. So we delete the ignore command in local test to make the ans files same in local and remote tests.

4. add test case 302 for gpload header and recover case 37 for fill_missing_fields