Update distribution spec for inner child of left outer index NL join to handle random distributions (6X) (#12303)
Previously, CPhysicalLeftOuterIndexNLJoin requested any distribution on the
inner child, and threw an exception (and fell back to planner) if the inner
table was randomly distributed to prevent incorrect results.  This commit adds
CPhysicalLeftOuterIndexNLJoin::EpetDistribution to prohibit certain plans so we
do not reach an invalid state (i.e. reject random distributions on the inner
child without creating a new distribution spec entirely).

This commit adds a specific case where `EpetDistribution` is called if we are
using `CPhysicalLeftOuterIndexNLJoin` (normally, an operation with
`CDistributionSpecAny` will not have `EpetDistribution` called, so we made an
exception specifc to this operation) and adds test cases to `gporca` to verify the
behavior is correct. This commit also adds missing test cases for hashed and
replicated tables with left outer index nested loop joins.