Generate additional partitioning request for greedy xform alternative in Orca
In CPhysicalInnerHashJoin::PppsRequired, we support 2 different
partitioning requests: 1 with DPE and 1 without DPE. Currently, we only
generate 1 partitioning request by default, but generate 2 if the join
order GUC is set to greedy/mincard/query. This was done to allow for
more alternatives without increasing the optimization time
significantly.

Now, we generate 2 requests for the greedy xform regardless of the GUC
setting. We saw that this helps generate some lower cost alternatives
with a small increase in optimization time (~8%). If we always
generated 2 requests, the optimization time would increase by a much
larger amount.

Some optimization time performance numbers for running the set of tpcds
queries with EXPLAIN (~100 queries) when `optimizer_join_order` is set to `exhaustive`:

tpcds base optimization time: 76s
tpcds optimization time with this change: 82s (~8% increase)

Co-authored-by: Chris Hajas <chajas@vmware.com>
Co-authored-by: Domino Valdano <dvaldano@vmware.com>