Compute SharedScan xslice buffer file names during ExecInitNode
ShareInputScan node uses temporary files as shared buffers and FIFOs as
a syncing mechanism for cross-slice communication (i.e IPC). The names
of these files is based on gp_session_id, gp_command_count & share_id.

Since these are computed (multiple times) during ExecProcNode, if the
gp_command_count (or another variable) changes between when the files
are created, written to and read from, the executor could ERROR failing
to find the required file, or even see wrong entries.

For the attached test case, ORCA produces a plan that executes an SPI
(pl/pgSQL) function on a master slice, which increments the gp_command
count, between the creation of the shared buffer file and it's
subsequent reading.

This commit fixes this issue by computing the names of all the files
required for cross-slice communication once during ExecInitNode() and
re-using that name for the rest of the query's execution.