Disallow unique rowid path when seeing indexonly scan.
Greenplum has a specical method to implement semjoin: it
might create unique rowid path to first do inner join and
then de-duplicated based on ctid. The method was added to
Greenplum long before but after merging so many commits
from upstream, some new logic might fail when applying this
method. IndexOnlyScan is a case that the de-dup logic did
not consider. IndexOnlyScan's var is pointing to the Index
Rel, not the Heap Rel, to totally fix the logic seems risky
on a stable branch like 6X. So this commit just disallows
the unique rowid method when seeing indexonly scan to avoid
error-out. Index paths are set during base rel, and planner
then generates join rels so that the disallow-loigc works.