Fix crash on QD when squelching CTE producer with a Motion child
Previously in commit 9fbd2da5635eecac6975e7349209373fc9177ef8, we added
special treatment for squelching share input scan by continueing
executing nodes below the CTE producer, so that a CTE consumer on a
differernt slice will not hang while waiting for the producer to finish
writing. It is also necessary for a squelched CTE producer on one
segment to finish fetching all the tuples and write them into a tuple
store, in case a cross-slice CTE consumer on a different segment does
actually need to read the real tuples.

However, there is an edge case: If an ERROR happend on QD before Shared
Scan being executed, the exectuor will directly abort the transaction
gracefully. In mppExecutorFinishup(), we first tear down the
interconnect, and if not all tuples are fetched, we call
ExecSquelchNode() after that. So, if there is a Motion below the CTE
producer, the Motion node will be executed during
ExecSquelchShareInputScan(), which will lead to QD crash because the
interconnect is already torn down.

Luckily, we do not need to finish executing the Motion in this edge
case, because we are aboring anyways. This patch adds an early return in
execMotionUnsortedReceiver() and execMotionSortedReceiver() for the
above case, and added ERROR handling if interconnect is down for other
reasons.

Note that we cannot simply first perform ExecSquelchNode() on QD and
then tear down its interconnect, because that would introduce deadlock
for some cases.

Reviewed-by: Shreedhar Hardikar <shardikar@vmware.com>
Reviewed-by: Adam Lee <adlee@vmware.com>
Reviewed-by: Zhenghua Lyu <zlyu@vmware.com>

(cherry picked from commit 38342426c07c0121659e8e8161bbc9b7beaa378a)