Fix assert check failed when a Subquery|function scan needs to create fake ctids in the update statement (#12511)
When the update statement contains semi-join of a subquery and others,  
we choose a unique_rowid_path for semi-join. But tuples from subquery
scan have no ctid.  When update statement contains a subquery, a WholeRowVar
is added to the target of the subquery, The slot is a virtual tuple in subqueryNext,
we assign ctid for this virtual tuple, however, After doing ExecTargetList for
WholeRowVar, this slot is formed to heap tuple, but the ctid of heap tuple is null.
when doing ExecTargetList for the target of ctid, we can not get ctid from heap
tuple. So the assert check failed.

Fix github issue: https://github.com/greenplum-db/gpdb/issues/12512

Co-authored-by: Zhenghua Lyu <kainwen@gmail.com>