Bitmap index replay test focused on meta-page not in shared_buffers
Test validates meta-page is read from disk during bitmpa LOV item
insert replay. There used to be bug such that if the metapage is not
present in shared buffers, it will not be fetched from disk. Instead,
a zeroed out page will be returned from memory. A subsequent flush of
the metapage will lead to an inadvertent permanent overwrite.

Reviewed-by: Soumyadeep Chakraborty <soumyadeep2007@gmail.com>