More robust implementation of MemoryContextContainsGenericAllocation
Instead of validating the memory just before the pointer we're
interested in, expecting it to contain a pointer to the context,
we look to see if the pointer is contained within any of the memory
blocks allocated by the memory context of interest.

This eliminates the possibility of a false positive happening due to
the contents of memory just before the pointer coincidentally matching one
of the shared header memory addresses.  It also eliminates false
negatives that happen when the pointer does not point to the beginning
of a chunk... resulting in an unnecessary mem copy.

Note:  this will only work for a MemoryContext which is an AllocSet.
To check whether a pointer is in an AllocSlab or a GenericSlab, we would
need to add separate handlers.  So far, these postgres features are not used
at all in gpdb.

Co-authored-by:  David Kimura <dkimura@vmware.com>