gprecoverseg: Add new wrappers for segment recovery
We have introduced two new wrappers named gpsegsetuprecovery and gpsegrecovery.
These will replace gpconfigurenewsegment which will now only be used by
gpexpand.
gpsegsetuprecovery: Takes care of datadir validation for full recovery and runs
the checkpoint query for incremental recovery.
gpsegsetuprecovery: Takes care of running either pg_rewind or pg_basebackup for
each required segment on that segment host.

Both of these files will be called from the coordinator to be run on a segment
host. On that segment host, each file will run all the operations in a pool for
all the required segments.

The caller will first call gpsegsetuprecovery in a pool and if something fails,
the execution will stop. Otherwise, the caller will call gpsegrecovery in a
pool to do the actual pg_rewind or pg_basebackup operation.

Co-authored-by: Divyesh Vanjare <vanjared@vmware.com>