Backport bitmap index scan when concurrent insert update full bitmap page on 6X_STABLE (#12731)
* Fix bitmap index scan when concurrent insert update full bitmap page

Concurrent scan on bitmap index when there's insert running in the
backend may cause the bitmap scan read wrong tid.
If a LOV has multiple bitmap pages, and the index insert tries to insert
a tid into a compressed word on a full bitmap page(Let's call the page
`PAGE_FULL`).
Then it'll try to find free space on next bitmap page(Let's call the
page `PAGE_NEXT`) and rearrange the words and copy extra words into
the next bitmap page.

So when the above insertion happens, imagine below case:
1. Query on bitmap: A query starts and reads all bitmap pages to
`PAGE_FULL`, increase
next tid to fetch, release lock after reading each page.
2. Concurrent insert: insert a tid into `PAGE_FULL` cause expand
compressed words to
new words, and rearrange words into `PAGE_NEXT`.
3. Query on bitmap: fetch `PAGE_NEXT` and expect the first tid in it
should equal the
saved next tid. But actually `PAGE_NEXT` now contains words used to
belong in `PAGE_FULL`.
This causes the real next tid less than the expected next tid. But our
scan keeps increasing
the wrong tid. And then this leads to a wrong result.
This related to issue: https://github.com/greenplum-db/gpdb/issues/11308

This commit done follow changes:
1. Allow Index Path on bitmap index for AO/AOCS tables. We used only
allow Bitmap Heap Path on AO/AOCS tables to avoid decompressed overhead
for random access data blocks. But Index Path on bitmap index will have
the same access order with Bitmap Heap Path.

2. BMBatchWords.firstTid is not well used in past. Now it always tracks
the first tid for the word to be checked for the current bitmap vector.

3. For BMIterateResult.nextTid, the init value for Bitmap Heap Scan is 1,
but for Index Scan it start from 0. Now make it consistent that always
start from 1. And it tracks the next tid to be checked during scanning.

4. When read BMBatchWords from bitmap page, we should calculate the
firstTid and nwordsread. We used to expect bitmap pages' tids are in
series. But this is not ture when the tid rearrange happens which
discussed above. The logic is in bitmapsearch.c `read_words`.

5. Consider catch up BMBatchWords.firstTid to BMIterateResult.nextTid,
so this could skip the scanned tids.

* Correct the real firstTid in _bitmap_union

In previous refactor which try to resolve issue:
https://github.com/greenplum-db/gpdb/issues/11308.
We change the way to calculate the firstTid holding in the
current position batch words, since concurrent insert will cause
the wrong tid matching in words_get_match().

To fix that issue in PR https://github.com/greenplum-db/gpdb/pull/11377,
we always get the correct firstTid from the bitmap index page.
But _bitmap_union is not touched. It was responsible to union bitmap
batch words for multivalues index scan like `where x in (x1, x2)` or `x > v`.

And also the logic for the catch up loop is not correct in
_bitmap_catchup_to_next_tid, it should check remain words number.
 
Co-authored-by: (Jerome)Junfeng Yang <yjerome@vmware.com>