Don't collect execution stats if the query failed
Currently, if one segment (like seg1) panics due to some reason, other
segments (like seg0) will report "InstrEndLoop called on running node".

This is because QE on seg1 reports ERROR to QD when execution fails,
and the QD dispatches cancel command to all the QEs, but another QE on
seg0 is not finished yet, its `instr` structure is not cleaned up.

This commit is to remove the logic of sending stats when execution
failed, which is not needed.

Fixes https://github.com/greenplum-db/gpdb/issues/9904

Signed-off-by: Adam Lee <adlee@vmware.com>