Set correct resultcollid for RelabelType
Earlier it was set to collation id of resulttype. But consider an expression
like  below :

string_to_array(unnest(anyarray)) <> '{nullout}'

annyarray has a collation oid of 0 but string_to_array returns a TEXTARRAY
which has a collation oid of 100.  So for resultcollid we should rely on
collation type of the expression.

Co-authored-by: Bhuvnesh Chaudhary bhuvnesh2703@gmail.com