gprecoverseg: Do not delete the cluster for newhost scenario
In order to test the `-p` functionality in gprecoverseg, we create a cluster
with spare hosts and then make one or more of the cluster hosts as unreachable
and use -p to make the segments on thost hosts failover to the spare hosts.
For this to work, we had to bring the cluster back to it's original state for
each example of the test scenario. This was done to make sure examples were not
dependent on each other and that they could start with a clean cluster. To
achieve this, we would delete the cluster using gpdeletesystem and reinitialize
it using gpinitsystem.
The problem with this approach was that any changes made to the cluster like
setting a guc would get removed because of running gpdeletesystem and
gpinitsystem. A better way is to reconnect the down hosts and then make the
segments on the spare hosts fallback to their original hosts.