ignore message '?' after nextval on QD is failed
when QE process nextval it sends NOTIFY to QD and asks QD to send nextval back to
QE, after getting nextval on QD is failed, QD send '?' message back to QE and cancel all
unfinished QEs, if the QE receives cancel before '?' message, QE will throw an error and
finish current command, the message '?' will stay in the socket, next time when we ReadCommand
we get '?' and ignore it. If we receive '?' before receiving the cancel signal, the next command will be OK,
it is flaky.