Exclude default gpbackup directory from pg_rewind
When running pg_rewind, things that exist on the target segment but
not on the source segment will be removed. In the case of HA when
recovering the preferred primary segment (mirror is active primary),
pg_rewind would try to delete the gpbackup default directory
"backups". Although this directory default is not recommended, it's a
valid path that needs to be excluded to protect backup data.

Note: This "backups" directory follows the same special GPDB exclusion
logic but with a unique difference. We must prevent the top-level
directory "backups" from being processed in the target files AND
exclude it in the source files exclusion check. This is because the
"backups" directory could exist only on the primary segment, only on
the mirror segment, or exist in both locations (which is different
from "pg_log" directory which always exists on both locations).

Backported from GPDB master branch:
https://github.com/greenplum-db/gpdb/commit/1a1863998eeb4ed097e2566383495909af0527ba

Merge conflict:
The pg_rewind TAP tests are not really used so had to translate the
TAP test change over to the regress-style pg_rewind extrafiles test.